+++ packages/stream_chat_flutter/lib/src/message_input/stream_message_input.dart
@@ -357,7 +357,6 @@ class StreamMessageInputState extends State<StreamMessageInput>
   late final CurvedAnimation _pickerAnimation;
 
   late StreamChatThemeData _streamChatTheme;
-  late StreamMessageInputThemeData _messageInputTheme;
 
   bool get _isEditing => _effectiveController.isEditing;
 
@@ -491,7 +490,6 @@ class StreamMessageInputState extends State<StreamMessageInput>
   @override
   void didChangeDependencies() {
     _streamChatTheme = StreamChatTheme.of(context);
-    _messageInputTheme = StreamMessageInputTheme.of(context);
     super.didChangeDependencies();
   }
 
@@ -583,7 +581,7 @@ class StreamMessageInputState extends State<StreamMessageInput>
     };
 
     final spacing = context.streamSpacing;
-    final safeAreaEnabled = widget.enableSafeArea ?? _messageInputTheme.enableSafeArea ?? true;
+    final safeAreaEnabled = widget.enableSafeArea ?? true;
     final viewPadding = MediaQuery.paddingOf(context);
 
     return Material(
@@ -741,13 +739,11 @@ class StreamMessageInputState extends State<StreamMessageInput>
 
     final allowedTypes = _getAllowedAttachmentPickerTypes();
 
-    final messageInputTheme = StreamMessageInputTheme.of(context);
     final isWebOrDesktop = switch (CurrentPlatform.type) {
       PlatformType.android || PlatformType.ios => false,
       _ => true,
     };
-    final useSystemPicker =
-        widget.useSystemAttachmentPicker || (messageInputTheme.useSystemAttachmentPicker ?? false) || isWebOrDesktop;
+    final useSystemPicker = widget.useSystemAttachmentPicker || isWebOrDesktop;
 
     final child = useSystemPicker
         ? systemAttachmentPickerBuilder(
@@ -790,7 +786,7 @@ class StreamMessageInputState extends State<StreamMessageInput>
       padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 15),
       child: Text(
         context.translations.sendMessagePermissionError,
-        style: _messageInputTheme.inputTextStyle,
+        style: context.streamTextInputTheme.style?.textStyle,
       ),
     );
   }