+++ packages/stream_chat_flutter/lib/src/theme/stream_chat_theme.dart
@@ -43,11 +43,7 @@ class StreamChatThemeData {
     StreamTextTheme? textTheme,
     StreamColorTheme? colorTheme,
     StreamChannelListHeaderThemeData? channelListHeaderTheme,
-    StreamChannelPreviewThemeData? channelPreviewTheme,
     StreamChannelHeaderThemeData? channelHeaderTheme,
-    StreamMessageThemeData? otherMessageTheme,
-    StreamMessageThemeData? ownMessageTheme,
-    StreamMessageInputThemeData? messageInputTheme,
     Widget Function(BuildContext, User)? defaultUserImage,
     PlaceholderUserImage? placeholderUserImage,
     IconThemeData? primaryIconTheme,
@@ -76,11 +72,7 @@ class StreamChatThemeData {
 
     final customizedData = defaultData.copyWith(
       channelListHeaderTheme: channelListHeaderTheme,
-      channelPreviewTheme: channelPreviewTheme,
       channelHeaderTheme: channelHeaderTheme,
-      otherMessageTheme: otherMessageTheme,
-      ownMessageTheme: ownMessageTheme,
-      messageInputTheme: messageInputTheme,
       defaultUserImage: defaultUserImage,
       placeholderUserImage: placeholderUserImage,
       primaryIconTheme: primaryIconTheme,
@@ -113,11 +105,7 @@ class StreamChatThemeData {
     required this.textTheme,
     required this.colorTheme,
     required this.channelListHeaderTheme,
-    required this.channelPreviewTheme,
     required this.channelHeaderTheme,
-    required this.otherMessageTheme,
-    required this.ownMessageTheme,
-    required this.messageInputTheme,
     required this.primaryIconTheme,
     required this.galleryHeaderTheme,
     required this.galleryFooterTheme,
@@ -151,7 +139,6 @@ class StreamChatThemeData {
     StreamColorTheme colorTheme,
     StreamTextTheme textTheme,
   ) {
-    final accentColor = colorTheme.accentPrimary;
     final iconTheme = IconThemeData(color: colorTheme.textLowEmphasis);
     final channelHeaderTheme = StreamChannelHeaderThemeData(
       avatarTheme: StreamAvatarThemeData(
@@ -163,31 +150,11 @@ class StreamChatThemeData {
       ),
       color: colorTheme.barsBg,
     );
-    final channelPreviewTheme = StreamChannelPreviewThemeData(
-      unreadCounterColor: colorTheme.accentError,
-      avatarTheme: StreamAvatarThemeData(
-        borderRadius: BorderRadius.circular(20),
-        constraints: const BoxConstraints.tightFor(
-          height: 40,
-          width: 40,
-        ),
-      ),
-      titleStyle: textTheme.bodyBold,
-      subtitleStyle: textTheme.footnote.copyWith(
-        color: const Color(0xff7A7A7A),
-      ),
-      lastMessageAtStyle: textTheme.footnote.copyWith(
-        // ignore: deprecated_member_use
-        color: colorTheme.textHighEmphasis.withOpacity(0.5),
-      ),
-      indicatorIconSize: 16,
-    );
 
     return StreamChatThemeData.raw(
       textTheme: textTheme,
       colorTheme: colorTheme,
       primaryIconTheme: iconTheme,
-      channelPreviewTheme: channelPreviewTheme,
       channelListHeaderTheme: StreamChannelListHeaderThemeData(
         avatarTheme: StreamAvatarThemeData(
           borderRadius: BorderRadius.circular(20),
@@ -200,93 +167,11 @@ class StreamChatThemeData {
         titleStyle: textTheme.headlineBold,
       ),
       channelHeaderTheme: channelHeaderTheme,
-      ownMessageTheme: StreamMessageThemeData(
-        messageAuthorStyle: textTheme.footnote.copyWith(color: colorTheme.textLowEmphasis),
-        messageTextStyle: textTheme.body,
-        messageDeletedStyle: textTheme.body.copyWith(
-          color: colorTheme.textLowEmphasis,
-          fontStyle: FontStyle.italic,
-        ),
-        createdAtStyle: textTheme.footnote.copyWith(color: colorTheme.textLowEmphasis),
-        repliesStyle: textTheme.footnoteBold.copyWith(color: accentColor),
-        messageBackgroundColor: colorTheme.inputBg,
-        reactionsBackgroundColor: colorTheme.barsBg,
-        reactionsBorderColor: colorTheme.borders,
-        reactionsMaskColor: colorTheme.appBg,
-        avatarTheme: StreamAvatarThemeData(
-          borderRadius: BorderRadius.circular(20),
-          constraints: const BoxConstraints.tightFor(
-            height: 32,
-            width: 32,
-          ),
-        ),
-        messageLinksStyle: TextStyle(color: accentColor),
-        urlAttachmentBackgroundColor: colorTheme.linkBg,
-        urlAttachmentHostStyle: textTheme.bodyBold.copyWith(color: accentColor),
-        urlAttachmentTitleStyle: textTheme.footnoteBold,
-        urlAttachmentTextStyle: textTheme.footnote,
-        urlAttachmentTitleMaxLine: 1,
-        urlAttachmentTextMaxLine: 3,
-      ),
-      otherMessageTheme: StreamMessageThemeData(
-        reactionsBackgroundColor: colorTheme.borders,
-        reactionsBorderColor: colorTheme.borders,
-        reactionsMaskColor: colorTheme.appBg,
-        messageTextStyle: textTheme.body,
-        messageDeletedStyle: textTheme.body.copyWith(
-          color: colorTheme.textLowEmphasis,
-          fontStyle: FontStyle.italic,
-        ),
-        createdAtStyle: textTheme.footnote.copyWith(color: colorTheme.textLowEmphasis),
-        messageAuthorStyle: textTheme.footnote.copyWith(color: colorTheme.textLowEmphasis),
-        repliesStyle: textTheme.footnoteBold.copyWith(color: accentColor),
-        messageLinksStyle: TextStyle(color: accentColor),
-        messageBackgroundColor: colorTheme.barsBg,
-        avatarTheme: StreamAvatarThemeData(
-          borderRadius: BorderRadius.circular(20),
-          constraints: const BoxConstraints.tightFor(
-            height: 32,
-            width: 32,
-          ),
-        ),
-        urlAttachmentBackgroundColor: colorTheme.linkBg,
-        urlAttachmentHostStyle: textTheme.bodyBold.copyWith(color: accentColor),
-        urlAttachmentTitleStyle: textTheme.footnoteBold,
-        urlAttachmentTextStyle: textTheme.footnote,
-        urlAttachmentTitleMaxLine: 1,
-        urlAttachmentTextMaxLine: 3,
-      ),
-      messageInputTheme: StreamMessageInputThemeData(
-        borderRadius: BorderRadius.circular(20),
-        sendAnimationDuration: const Duration(milliseconds: 300),
-        actionButtonColor: colorTheme.accentPrimary,
-        actionButtonIdleColor: colorTheme.textLowEmphasis,
-        expandButtonColor: colorTheme.accentPrimary,
-        sendButtonColor: colorTheme.accentPrimary,
-        sendButtonIdleColor: colorTheme.disabled,
-        inputBackgroundColor: colorTheme.barsBg,
-        inputTextStyle: textTheme.body,
-        linkHighlightColor: colorTheme.accentPrimary,
-        idleBorderGradient: LinearGradient(
-          colors: [
-            colorTheme.borders,
-            colorTheme.borders,
-          ],
-        ),
-        activeBorderGradient: LinearGradient(
-          colors: [
-            colorTheme.borders,
-            colorTheme.borders,
-          ],
-        ),
-        useSystemAttachmentPicker: false,
-      ),
       galleryHeaderTheme: StreamGalleryHeaderThemeData(
         closeButtonColor: colorTheme.textHighEmphasis,
         backgroundColor: channelHeaderTheme.color,
         iconMenuPointColor: colorTheme.textHighEmphasis,
         titleTextStyle: textTheme.headlineBold,
-        subtitleTextStyle: channelPreviewTheme.subtitleStyle,
         bottomSheetBarrierColor: colorTheme.overlay,
       ),
       galleryFooterTheme: StreamGalleryFooterThemeData(
@@ -429,9 +314,6 @@ class StreamChatThemeData {
   /// The color themes used in the widgets
   final StreamColorTheme colorTheme;
 
-  /// Theme of the [StreamChannelPreview]
-  final StreamChannelPreviewThemeData channelPreviewTheme;
-
   /// Theme of the [StreamChannelListHeader]
   final StreamChannelListHeaderThemeData channelListHeaderTheme;
 
@@ -445,15 +327,6 @@ class StreamChatThemeData {
   /// The default style for [StreamGalleryFooter]s below the overall
   /// [StreamChatTheme].
   final StreamGalleryFooterThemeData galleryFooterTheme;
-
-  /// Theme of the current user messages
-  final StreamMessageThemeData ownMessageTheme;
-
-  /// Theme of other users messages
-  final StreamMessageThemeData otherMessageTheme;
-
-  /// Theme dedicated to the [StreamMessageInput] widget
-  final StreamMessageInputThemeData messageInputTheme;
 
   /// Primary icon theme
   final IconThemeData primaryIconTheme;
@@ -491,25 +364,12 @@ class StreamChatThemeData {
   /// Theme configuration for the [StreamDraftListTile] widget.
   final StreamDraftListTileThemeData draftListTileTheme;
 
-  /// Returns the theme for the message based on the [reverse] parameter.
-  ///
-  /// If [reverse] is true, it returns the [otherMessageTheme], otherwise it
-  /// returns the [ownMessageTheme].
-  StreamMessageThemeData getMessageTheme({bool reverse = false}) {
-    if (reverse) return ownMessageTheme;
-    return otherMessageTheme;
-  }
-
   /// Creates a copy of [StreamChatThemeData] with specified attributes
   /// overridden.
   StreamChatThemeData copyWith({
     StreamTextTheme? textTheme,
     StreamColorTheme? colorTheme,
-    StreamChannelPreviewThemeData? channelPreviewTheme,
     StreamChannelHeaderThemeData? channelHeaderTheme,
-    StreamMessageThemeData? ownMessageTheme,
-    StreamMessageThemeData? otherMessageTheme,
-    StreamMessageInputThemeData? messageInputTheme,
     Widget Function(BuildContext, User)? defaultUserImage,
     PlaceholderUserImage? placeholderUserImage,
     IconThemeData? primaryIconTheme,
@@ -532,11 +392,7 @@ class StreamChatThemeData {
     textTheme: this.textTheme.merge(textTheme),
     colorTheme: this.colorTheme.merge(colorTheme),
     primaryIconTheme: this.primaryIconTheme.merge(primaryIconTheme),
-    channelPreviewTheme: this.channelPreviewTheme.merge(channelPreviewTheme),
     channelHeaderTheme: this.channelHeaderTheme.merge(channelHeaderTheme),
-    ownMessageTheme: this.ownMessageTheme.merge(ownMessageTheme),
-    otherMessageTheme: this.otherMessageTheme.merge(otherMessageTheme),
-    messageInputTheme: this.messageInputTheme.merge(messageInputTheme),
     galleryHeaderTheme: galleryHeaderTheme ?? this.galleryHeaderTheme,
     galleryFooterTheme: galleryFooterTheme ?? this.galleryFooterTheme,
     messageListViewTheme: messageListViewTheme ?? this.messageListViewTheme,
@@ -560,11 +416,7 @@ class StreamChatThemeData {
       textTheme: textTheme.merge(other.textTheme),
       colorTheme: colorTheme.merge(other.colorTheme),
       primaryIconTheme: other.primaryIconTheme,
-      channelPreviewTheme: channelPreviewTheme.merge(other.channelPreviewTheme),
       channelHeaderTheme: channelHeaderTheme.merge(other.channelHeaderTheme),
-      ownMessageTheme: ownMessageTheme.merge(other.ownMessageTheme),
-      otherMessageTheme: otherMessageTheme.merge(other.otherMessageTheme),
-      messageInputTheme: messageInputTheme.merge(other.messageInputTheme),
       galleryHeaderTheme: galleryHeaderTheme.merge(other.galleryHeaderTheme),
       galleryFooterTheme: galleryFooterTheme.merge(other.galleryFooterTheme),
       messageListViewTheme: messageListViewTheme.merge(other.messageListViewTheme),